Implementation of PathLock features.
To manage concurrency this several classes has been added:

- PathLock which is a lock associated to a Path and
which can be locked for read or for write. This class
implement AutoClosable to allow the use of the lock
in a "Try with ressources" statement.

- FuseLockManager which allows to retrieve PathLock
from Path.

- BasicFuseLockManager which is an implementation
of FuseLockManager and which retrieve PathLock
by using Google Guava Cache.

Following the add of these classes, update
ConcreteFuseFS by adding the use of PathLock
where it is needed.