Fix PathLock bug.
PathLock try to unlock read and write lock
in the same time. However, it throws an
IllegalMonitorStateException because
unlock() methods can be called only on
lock that we previously take.

To fix it. Remove Autocloseable from PathLock
and create CloseableLock which is a Lock
which can be closed in a "Try with resources"
statement.

Update ConcreteFuseFS in consequences.