Create runnable examples and fix bugs
Create "example" package which contains
GuiceModule and Main class to run an
example with NIO Fuse.

Fix get attr bug

Get attr do not correctly set access to
file on Windows (not tested on linux).

Access has been review and now write
actions can be done and work correctly

Fix read bug

BasicFileReference has a bug.
When EOF was reach, the file channel
returns -1. We use comparaison to know
if the EOF was reach however the condition
was wrong, instead of read == -1, it was
wrote read != -1.

In addition, if the EOF has been reach,
put in the buf pointer the rest of the
buffer which has been read.

Bug has been fixed

Fix bug of OneDirectoryLinkPathRecover

FusePath begin with "/" or "\" and this make
that the method Path.resolve() returns a
not correct value.
To fix that, remove the first character "/"
or "\" of the fusePath given in the
OneDirectoryLinkPathRecover.recover() method