Update NIOFuseDirectoryManager by changing the way to filter file in directory.
Create interface DirectoryFileFilter which propose a method
which determine if a file of a directory can be accepted or
not.

A DirectoryFileFilter is now injected in NIOFuseDirectoryManager
and the NIOFuseDirectoryManager use it to filter file which
must displayed or not in the Fuse Directory. The dependency
can be null.

Update NIOGuiceModule to add DirectoryFileFilter dependency.